L2e + L2d: every cell carries a reading, and spelling moves home
Two stages of L2 (#229), plus a finding that a third does not exist.

L2e -- TypeSubject is deleted.

The enum had two variants: a type the flat API contains, and "a type only
the binding authored", which carried no reading on the theory that a
declared wire type or an unfold leaf had none to give. It did. Those are
ordinary types in this language; they were only absent from an index of what
the SOURCE wrote. L2a made ensure_entry admit them, and the re-measurement
promised there found ZERO grammar refusals across every in-tree example and
all 523 tests -- so the second variant had no members left.

TypeCell::subject is now the TypeRef itself. A spelling the grammar really
does refuse is a ScanError::NotExpressible naming it, rather than a cell that
quietly means less than its neighbours: ensure_entry returns Result, and
every caller already did. Only an entry point can reach that error at all --
a type the walk found came from an existing reading's origin.syntax, so it
lowered once already.

location() and kind() go with the enum. The has_position() gate moves to the
two resolve.rs sites that were its only production readers, unchanged: a
reading and a reportable position stay two facts, which is the whole reason
a binding-local sig!(..) does not report `:0:0:`.

L2d -- spelling moves into core::flat.

type_from_ident, canonical_type, canonical_spelling, normalize_type,
normalize_item_types, reduce_flat_path, constructor_key and Normalization
become core::flat::spelling. They decide what spelling a type HAS before
anything keys on it, which is the same authority that decides what a type
MEANS, and they were sitting in types_util next to the classifiers looking
like one of them.

Checked rather than assumed: all eight have ZERO callers outside api/core,
and normalize_* is called only by the frontend itself. That is what makes
them movable while the rest of types_util waits on L3/L4 -- option_inner_type
has 40 adapter callers, bare_path_ident 22, is_unit 18.

A move, not a migration, and labelled as one: the four ledger sites leave
api/core because the code left, exactly as #248's did. types_util 14 -> 10.

L2c does not exist.

api/core has ZERO production callers of enum_shape, first_payload_variant or
enum_discriminant_values -- the three I counted when planning were all in
types_util/tests.rs, which my grep excluded types_util.rs but not its test
module. Every real caller is in an adapter, so the whole stage is L3/L4 work
and there is nothing here to do.

L2b is NOT in this PR.

It is 20 sites, and only about half are the Option/Vec/ref fold that a layer
read on TypeRef replaces. The rest ask different questions -- a slice element,
a callback argument, is_nominated on a peeled element -- and migrating the
easy half would leave unfold.rs holding both idioms, which is worse than
either and moves the ledger barely at all. It gets its own PR, with the
helper's shape driven by all 20 sites rather than the first two. I wrote that
helper, found it did not survive contact with the harder sites, and removed
it rather than ship an unused public API.

Ledger 158 -> 154.

Reported: regen-check is byte-identical on every committed artifact. It also
reports two UNTRACKED example_flat_aarch64_unstable.{rs,h} files -- verified
identical on language-integration with the same clean+rebuild, so that is the
pre-existing #219 disagreement, not this change.
Explained: no generated output moved, and none should have -- both stages are
internal.
Asserted: every type-table cell carries the frontend's reading of its type,
and there is no second kind that carries nothing.
