+++ prebindgen/src/api/core/flat/boundary.ledger
@@ -46,7 +46,7 @@
 
 4	api/core/expand.rs
 2	api/core/registry/scan.rs
-14	api/core/types_util.rs
+10	api/core/types_util.rs
 16	api/core/unfold.rs
 8	api/lang/cbindgen/builder.rs
 1	api/lang/cbindgen/convert.rs
@@ -73,4 +73,4 @@
 2	api/lang/jnigen/jni/wire_access.rs
 2	api/lang/jnigen/util.rs
 
-# total: 158
+# total: 154
